backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.4
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-js-asset==3.0.1
djangorestframework==3.15.2
fingerprint-pro-server-api-sdk==8.1.0
idna==3.10
pillow==11.0.0
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
